feat: refactored how default optin template is rendered.
This is more aligned to how I imagine a WordPress developer would want to initialize each part of the library.

I'm using an add_action in the Plugin implementation instead of embedding it into the run_optin because we don't know which action the dev needs to use for their needs.

I added add_action calls in the enqueue function of the default optin template class because I'm letting the developer decide how to enqueue their styles and scripts. If they don't need to enqueue an actual file, they can just render HTML code like I did.