Avoid recommending strauss_overload (#59)
The strauss_overload messes with the namespace overrides that Strauss is meant to perform. Instead, to force Strauss to copy/address directories, you can set up fake namespaces for PSR-4 autoloading and that causes Strauss to properly copy the files. In this library's case, it copies `src/Telemetry`, `src/views`, and `src/resources`.